robot-ai/requirements.txt
ultralytics>=8.0.0
torch>=2.0.0
torchvision
torchaudio
stable-baselines3[extra]>=2.0.0
gym==0.26.2
opencv-python-headless
numpy
pandas
matplotlib
transformers>=4.40.0
accelerate
pillow
roboflow
requests
fastapi
uvicorn[standard]
python-multipart
supervision
